requirements.txt
fastapi
uvicorn[standard]
langgraph
yfinance
pandas
numpy
requests
python-dotenv
google-genai
